chore: remove no longer needed integration funcs (#1109)
