+++ integrations/aws/client.go
@@ -2,28 +2,15 @@ package aws
 
 import (
 	"context"
-	"os"
-	"strconv"
 
-	"github.com/aws/aws-sdk-go-v2/aws"
 	"github.com/aws/aws-sdk-go-v2/config"
 	"github.com/aws/aws-sdk-go-v2/credentials"
-	"github.com/aws/aws-sdk-go-v2/service/ec2"
-	"github.com/aws/aws-sdk-go-v2/service/eventbridge"
-	"github.com/aws/aws-sdk-go-v2/service/iam"
-	"github.com/aws/aws-sdk-go-v2/service/rds"
-	"github.com/aws/aws-sdk-go-v2/service/rdsdata"
-	"github.com/aws/aws-sdk-go-v2/service/s3"
-	"github.com/aws/aws-sdk-go-v2/service/sns"
-	"github.com/aws/aws-sdk-go-v2/service/sqs"
 	"github.com/aws/aws-sdk-go-v2/service/sts"
 	"go.uber.org/zap"
 
 	"go.autokitteh.dev/autokitteh/integrations"
 	"go.autokitteh.dev/autokitteh/integrations/common"
-	"go.autokitteh.dev/autokitteh/internal/kittehs"
 	"go.autokitteh.dev/autokitteh/sdk/sdkintegrations"
-	"go.autokitteh.dev/autokitteh/sdk/sdklogger"
 	"go.autokitteh.dev/autokitteh/sdk/sdkmodule"
 	"go.autokitteh.dev/autokitteh/sdk/sdkservices"
 	"go.autokitteh.dev/autokitteh/sdk/sdktypes"
@@ -31,64 +18,15 @@ import (
 
 var desc = common.Descriptor("aws", "AWS (All APIs)", "/static/images/aws.svg")
 
-type integration struct {
-	vars sdkservices.Vars
-}
+type integration struct{ vars sdkservices.Vars }
 
-var (
-	svcs = []struct {
-		name string
-		fn   any
-	}{
-		{"ec2", ec2.NewFromConfig},
-		{"eventbridge", eventbridge.NewFromConfig},
-		{"iam", iam.NewFromConfig},
-		{"rds", rds.NewFromConfig},
-		{"rdsdata", rdsdata.NewFromConfig},
-		{"s3", s3.NewFromConfig},
-		{"sns", sns.NewFromConfig},
-		{"sqs", sqs.NewFromConfig},
-	}
-
-	useDefaultConfig, _ = strconv.ParseBool(os.Getenv("AWS_USE_DEFAULT_CONFIG"))
-
-	defaultAWSConfig *aws.Config
-
-	authType = sdktypes.NewSymbol("authType")
-)
-
-func init() {
-	initDefaultConfig()
-}
-
-func initDefaultConfig() {
-	if !useDefaultConfig {
-		return
-	}
-
-	sdklogger.Warn("aws: using default global config")
-
-	cfg, err := config.LoadDefaultConfig(context.Background())
-	if err != nil {
-		sdklogger.Panic(err)
-		return
-	}
-
-	defaultAWSConfig = &cfg
-}
-
-func initOpts(vars sdkservices.Vars) (opts []sdkmodule.Optfn) {
-	for _, svc := range svcs {
-		opts = append(opts, kittehs.Must1(importServiceMethods(vars, svc.name, svc.fn))...)
-	}
-	return
-}
+var authType = sdktypes.NewSymbol("authType")
 
 func New(cvars sdkservices.Vars) sdkservices.Integration {
 	i := &integration{vars: cvars}
 	return sdkintegrations.NewIntegration(
 		desc,
-		sdkmodule.New(initOpts(cvars)...),
+		sdkmodule.New(),
 		connStatus(i),
 		connTest(i),
 		sdkintegrations.WithConnectionConfigFromVars(cvars),
