+++ integrations/chatgpt/client.go
@@ -27,14 +27,7 @@ func New(vars sdkservices.Vars) sdkservices.Integration {
 	i := &integration{vars: vars}
 	return sdkintegrations.NewIntegration(
 		desc,
-		sdkmodule.New(
-			sdkmodule.ExportFunction(
-				"create_chat_completion",
-				i.createChatCompletion,
-				sdkmodule.WithFuncDoc("https://pkg.go.dev/github.com/sashabaranov/go-openai#Client.CreateChatCompletion"),
-				sdkmodule.WithArgs("model?", "message?", "messages?"),
-			),
-		),
+		sdkmodule.New(),
 		connStatus(i),
 		connTest(i),
 		sdkintegrations.WithConnectionConfigFromVars(vars),
