+++ integrations/twilio/client.go
@@ -24,13 +24,7 @@ func New(vars sdkservices.Vars) sdkservices.Integration {
 
 	return sdkintegrations.NewIntegration(
 		desc,
-		sdkmodule.New(
-			sdkmodule.ExportFunction("create_message",
-				i.createMessage,
-				sdkmodule.WithFuncDoc("https://www.twilio.com/docs/messaging/api/message-resource#create-a-message-resource"),
-				sdkmodule.WithArgs("to", "from_number?", "messaging_service_sid?", "body?", "media_url?", "content_sid?"),
-			),
-		),
+		sdkmodule.New(),
 		connStatus(i),
 		connTest(i),
 		sdkintegrations.WithConnectionConfigFromVars(vars),
