+++ internal/backend/sessions/sessionworkflows/modules/akmodules.go
@@ -4,7 +4,6 @@ import (
 	"go.autokitteh.dev/autokitteh/internal/backend/fixtures"
 	"go.autokitteh.dev/autokitteh/internal/backend/sessions/sessionworkflows/modules/http"
 	"go.autokitteh.dev/autokitteh/internal/backend/sessions/sessionworkflows/modules/os"
-	"go.autokitteh.dev/autokitteh/internal/backend/sessions/sessionworkflows/modules/store"
 	"go.autokitteh.dev/autokitteh/internal/backend/sessions/sessionworkflows/modules/testtools"
 	"go.autokitteh.dev/autokitteh/internal/backend/sessions/sessionworkflows/modules/time"
 	"go.autokitteh.dev/autokitteh/sdk/sdktypes"
@@ -12,7 +11,7 @@ import (
 
 func IsAKModuleExecutorID(xid sdktypes.ExecutorID) bool {
 	switch xid {
-	case store.ExecutorID, time.ExecutorID, os.ExecutorID, fixtures.ModuleExecutorID, http.ExecutorID, testtools.ExecutorID:
+	case time.ExecutorID, os.ExecutorID, fixtures.ModuleExecutorID, http.ExecutorID, testtools.ExecutorID:
 		return true
 	default:
 		return false
