+++ internal/backend/sessions/sessionworkflows/workflow.go
@@ -19,7 +19,6 @@ import (
 	"go.autokitteh.dev/autokitteh/internal/backend/sessions/sessiondata"
 	httpmodule "go.autokitteh.dev/autokitteh/internal/backend/sessions/sessionworkflows/modules/http"
 	osmodule "go.autokitteh.dev/autokitteh/internal/backend/sessions/sessionworkflows/modules/os"
-	"go.autokitteh.dev/autokitteh/internal/backend/sessions/sessionworkflows/modules/store"
 	testtoolsmodule "go.autokitteh.dev/autokitteh/internal/backend/sessions/sessionworkflows/modules/testtools"
 	timemodule "go.autokitteh.dev/autokitteh/internal/backend/sessions/sessionworkflows/modules/time"
 	"go.autokitteh.dev/autokitteh/internal/backend/temporalclient"
@@ -258,10 +257,9 @@ func (w *sessionWorkflow) initConnections(wctx workflow.Context) (map[string]con
 
 func (w *sessionWorkflow) initGlobalModules() (map[string]sdktypes.Value, error) {
 	execs := map[string]sdkexecutor.Executor{
-		"ak":    w.newModule(),
-		"time":  timemodule.New(),
-		"http":  httpmodule.New(),
-		"store": store.New(w.data.Session.ProjectID(), w.ws.svcs.RedisClient),
+		"ak":   w.newModule(),
+		"time": timemodule.New(),
+		"http": httpmodule.New(),
 	}
 
 	vs := make(map[string]sdktypes.Value, len(execs))
